game_of_life example
