+++ examples/game_of_life/assets/favicon.ico
