+++ examples/game_of_life/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <title>Yew • Game of Life</title>
+
+    <link rel="stylesheet" href="styles.css" />
+    <link rel="trunk-dist" href="assets/favicon.ico" />
+  </head>
+
+  <body></body>
+</html>
