+++ examples/game_of_life/src/main.rs
@@ -1,13 +1,236 @@
-use game_of_life::{Model, Msg};
-use log::trace;
-use yew::App;
+use cell::Cellule;
+use rand::Rng;
+use std::time::Duration;
+use yew::services::interval::{IntervalService, IntervalTask};
+use yew::{html, Component, ComponentLink, Html, ShouldRender};
+
+mod cell;
+
+pub enum Msg {
+    Random,
+    Start,
+    Step,
+    Reset,
+    Stop,
+    ToggleCellule(usize),
+    Tick,
+}
+
+pub struct Model {
+    link: ComponentLink<Self>,
+    active: bool,
+    cellules: Vec<Cellule>,
+    cellules_width: usize,
+    cellules_height: usize,
+    _task: IntervalTask,
+}
+
+impl Model {
+    pub fn random_mutate(&mut self) {
+        for cellule in self.cellules.iter_mut() {
+            if rand::thread_rng().gen() {
+                cellule.set_alive();
+            } else {
+                cellule.set_dead();
+            }
+        }
+    }
+
+    fn reset(&mut self) {
+        for cellule in self.cellules.iter_mut() {
+            cellule.set_dead();
+        }
+    }
+
+    fn step(&mut self) {
+        let mut to_dead = Vec::new();
+        let mut to_live = Vec::new();
+        for row in 0..self.cellules_height {
+            for col in 0..self.cellules_width {
+                let neighbors = self.neighbors(row as isize, col as isize);
+
+                let current_idx = self.row_col_as_idx(row as isize, col as isize);
+                if self.cellules[current_idx].is_alive() {
+                    if Cellule::alone(&neighbors) || Cellule::overpopulated(&neighbors) {
+                        to_dead.push(current_idx);
+                    }
+                } else if Cellule::can_be_revived(&neighbors) {
+                    to_live.push(current_idx);
+                }
+            }
+        }
+        to_dead
+            .iter()
+            .for_each(|idx| self.cellules[*idx].set_dead());
+        to_live
+            .iter()
+            .for_each(|idx| self.cellules[*idx].set_alive());
+    }
+
+    fn neighbors(&self, row: isize, col: isize) -> [Cellule; 8] {
+        [
+            self.cellules[self.row_col_as_idx(row + 1, col)],
+            self.cellules[self.row_col_as_idx(row + 1, col + 1)],
+            self.cellules[self.row_col_as_idx(row + 1, col - 1)],
+            self.cellules[self.row_col_as_idx(row - 1, col)],
+            self.cellules[self.row_col_as_idx(row - 1, col + 1)],
+            self.cellules[self.row_col_as_idx(row - 1, col - 1)],
+            self.cellules[self.row_col_as_idx(row, col - 1)],
+            self.cellules[self.row_col_as_idx(row, col + 1)],
+        ]
+    }
+
+    fn row_col_as_idx(&self, row: isize, col: isize) -> usize {
+        let row = wrap(row, self.cellules_height as isize);
+        let col = wrap(col, self.cellules_width as isize);
+
+        row * self.cellules_width + col
+    }
+
+    fn view_cellule(&self, idx: usize, cellule: &Cellule) -> Html {
+        let cellule_status = {
+            if cellule.is_alive() {
+                "cellule-live"
+            } else {
+                "cellule-dead"
+            }
+        };
+        html! {
+            <div key=idx class=("game-cellule", cellule_status)
+                onclick=self.link.callback(move |_| Msg::ToggleCellule(idx))>
+            </div>
+        }
+    }
+}
+impl Component for Model {
+    type Message = Msg;
+    type Properties = ();
+
+    fn create(_props: Self::Properties, link: ComponentLink<Self>) -> Self {
+        let callback = link.callback(|_| Msg::Tick);
+        let task = IntervalService::spawn(Duration::from_millis(200), callback);
+
+        let (cellules_width, cellules_height) = (53, 40);
+
+        Self {
+            link,
+            active: false,
+            cellules: vec![Cellule::new_dead(); cellules_width * cellules_height],
+            cellules_width,
+            cellules_height,
+            _task: task,
+        }
+    }
+
+    fn update(&mut self, msg: Self::Message) -> ShouldRender {
+        match msg {
+            Msg::Random => {
+                self.random_mutate();
+                log::info!("Random");
+                true
+            }
+            Msg::Start => {
+                self.active = true;
+                log::info!("Start");
+                false
+            }
+            Msg::Step => {
+                self.step();
+                true
+            }
+            Msg::Reset => {
+                self.reset();
+                log::info!("Reset");
+                true
+            }
+            Msg::Stop => {
+                self.active = false;
+                log::info!("Stop");
+                false
+            }
+            Msg::ToggleCellule(idx) => {
+                let cellule = self.cellules.get_mut(idx).unwrap();
+                cellule.toggle();
+                true
+            }
+            Msg::Tick => {
+                if self.active {
+                    self.step();
+                    true
+                } else {
+                    false
+                }
+            }
+        }
+    }
+
+    fn change(&mut self, _props: Self::Properties) -> ShouldRender {
+        unimplemented!()
+    }
+
+    fn view(&self) -> Html {
+        let cell_rows =
+            self.cellules
+                .chunks(self.cellules_width)
+                .enumerate()
+                .map(|(y, cellules)| {
+                    let idx_offset = y * self.cellules_width;
+
+                    let cells = cellules
+                        .iter()
+                        .enumerate()
+                        .map(|(x, cell)| self.view_cellule(idx_offset + x, cell));
+                    html! {
+                        <div key=y class="game-row">
+                            { for cells }
+                        </div>
+                    }
+                });
+
+        html! {
+            <div>
+                <section class="game-container">
+                    <header class="app-header">
+                        <img src="favicon.ico" class="app-logo"/>
+                        <h1 class="app-title">{ "Game of Life" }</h1>
+                    </header>
+                    <section class="game-area">
+                        <div class="game-of-life">
+                            { for cell_rows }
+                        </div>
+                        <div class="game-buttons">
+                            <button class="game-button" onclick=self.link.callback(|_| Msg::Random)>{ "Random" }</button>
+                            <button class="game-button" onclick=self.link.callback(|_| Msg::Step)>{ "Step" }</button>
+                            <button class="game-button" onclick=self.link.callback(|_| Msg::Start)>{ "Start" }</button>
+                            <button class="game-button" onclick=self.link.callback(|_| Msg::Stop)>{ "Stop" }</button>
+                            <button class="game-button" onclick=self.link.callback(|_| Msg::Reset)>{ "Reset" }</button>
+                        </div>
+                    </section>
+                </section>
+                <footer class="app-footer">
+                    <strong class="footer-text">
+                      { "Game of Life - a yew experiment " }
+                    </strong>
+                    <a href="https://github.com/yewstack/yew" target="_blank">{ "source" }</a>
+                </footer>
+            </div>
+        }
+    }
+}
+
+fn wrap(coord: isize, range: isize) -> usize {
+    let result = if coord < 0 {
+        coord + range
+    } else if coord >= range {
+        coord - range
+    } else {
+        coord
+    };
+    result as usize
+}
 
 fn main() {
     wasm_logger::init(wasm_logger::Config::default());
-    trace!("Initializing yew...");
-    yew::initialize();
-    App::<Model>::new()
-        .mount_to_body()
-        .send_message(Msg::Random);
-    yew::run_loop();
+    log::trace!("Initializing yew...");
+    yew::start_app::<Model>();
 }
