--- examples/game_of_life/static/index.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <title>Yew • Game of Life</title>
-        <link rel="stylesheet" href="./styles.css">
-        <script type="module">
-            import init from "./wasm.js"
-            init()
-        </script>
-    </head>
-    <body></body>
-</html>
