+++ examples/game_of_life/styles.css
